[LGgenFDI] Fix @smejl comments
